requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

sqlalchemy==2.0.32
pydantic==2.9.2

python-multipart==0.0.9

passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

pypdf==4.2.0
